test: add test coverage :100:
